backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.9
sqlparse==0.5.3
python-dotenv==1.0.1
gunicorn==22.0.0
django-cors-headers==4.3.1
